+++ src/classes/board/ball.cpp
@@ -1,6 +1,6 @@
 #include "ball.h"
-#include "Vec2f.h"
-#include "utils_2d.h"
+#include "../Vec2f.h"
+#include "../utils_2d.h"
 #include "wall.h"
 #include <SFML/Graphics/CircleShape.hpp>
 #include <SFML/Graphics/Color.hpp>
