+++ src/classes/board/chargeMeter.h
@@ -1,7 +1,7 @@
 #pragma once
 
+#include "../Vec2f.h"
 #include "ball.h"
-#include "Vec2f.h"
 #include <SFML/Graphics/RenderWindow.hpp>
 #include <SFML/Graphics/VertexArray.hpp>
