+++ src/classes/gui/button.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Vec2f.h"
+#include "../Vec2f.h"
 #include <SFML/Graphics/RectangleShape.hpp>
 #include <SFML/Graphics/RenderWindow.hpp>
 #include <SFML/Graphics/Text.hpp>
