+++ src/classes/gui/gui.cpp
@@ -1,7 +1,7 @@
 #include "gui.h"
+#include "../Vec2f.h"
 #include "SFML/Graphics/RenderWindow.hpp"
 #include "button.h"
-#include "Vec2f.h"
 #include "slider.h"
 #include <utility>
 #include <vector>
